refactor: 의존성 UserRepository로 변경

+++ backend/src/main/java/com/team5/issue_tracker/milestone/service/MilestoneService.java
@@ -33,6 +33,9 @@ public class MilestoneService {
 
   @Transactional
   public void deleteMilestoneById(Long id) {
+    if (!milestoneRepository.existsById(id)) {
+      throw new NotFoundException(ErrorCode.MILESTONE_NOT_FOUND);
+    }
     milestoneRepository.deleteById(id);
   }
